refactor ARPlacePlantView
